+++ app/presenters/currency_answer_presenter.rb
@@ -1,0 +1,7 @@
+class CurrencyAnswerPresenter < SimpleDelegator
+  include ActionView::Helpers::NumberHelper
+
+  def response
+    number_to_currency(super, unit: "£", precision: 2)
+  end
+end